+++ framework/model-test/src/main/java/is/codion/framework/model/test/AbstractEntityModelTest.java
@@ -104,7 +104,10 @@ public abstract class AbstractEntityModelTest<Model extends DefaultEntityModel<M
 
   @Test
   public void testDetailModels() throws DatabaseException, ValidationException {
-    //todo
+    Model departmentModel = createDepartmentModel();
+    assertEquals(1, departmentModel.activeDetailModels().size());
+    departmentModel.detailModelLink(departmentModel.detailModel(Employee.TYPE)).active().set(false);
+    assertTrue(departmentModel.activeDetailModels().empty());
   }
 
   @Test
@@ -205,9 +208,9 @@ public abstract class AbstractEntityModelTest<Model extends DefaultEntityModel<M
   @Test
   public void activateDeactivateDetailModel() {
     departmentModel.detailModelLink(departmentModel.detailModel(Employee.TYPE)).active().set(false);
-    assertTrue(departmentModel.activeDetailModels().isEmpty());
+    assertTrue(departmentModel.activeDetailModels().get().isEmpty());
     departmentModel.detailModelLink(departmentModel.detailModel(Employee.TYPE)).active().set(true);
-    assertFalse(departmentModel.activeDetailModels().isEmpty());
+    assertFalse(departmentModel.activeDetailModels().empty());
     assertTrue(departmentModel.activeDetailModels().contains(departmentModel.detailModel(Employee.TYPE)));
   }