+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/TabbedDetailLayout.java
@@ -424,7 +424,7 @@ public final class TabbedDetailLayout implements DetailLayout {
     private void activateDetailModelLink(SwingEntityModel detailModel) {
       SwingEntityModel model = entityPanel.model();
       if (model.containsDetailModel(detailModel)) {
-        model.activeDetailModels().stream()
+        model.activeDetailModels().get().stream()
                 .filter(activeDetailModel -> activeDetailModel != detailModel)
                 .forEach(activeDetailModel -> model.detailModelLink(activeDetailModel).active().set(false));
         model.detailModelLink(detailModel).active().set(true);