+++ changelog.md
@@ -8,6 +8,8 @@ Codion Change Log
 - ValueSetObserver.size() added.
 ### is.codion.framework.i18n
 - FrameworkMessages.confirmDeleteSelected() removed.
+### is.codion.swing.framework.model
+- EntityModel.activeDetailModels() now returns ValueSetObserver, addActiveDetailModelsListener() and removeActiveDetailModelsListener() removed.
 ### is.codion.swing.framework.ui
 - EntityPanel.detailLayout() now protected instead of package private.
 - EntityPanel.selectDetailPanel() removed for now.